requirements.txt
fastapi==0.75.2
nltk==3.7
numpy==1.22.3
tensorflow-cpu==2.8.0
uvicorn==0.17.6
gunicorn
